Move DocCommentConverter to Microsoft.CodeAnalysis.CSharp.DocumentationComments
